Reworked cube class to allow construction of any size of cubes (spoiler: only 3x3 version works currently), working towards a generic cube-turning function 'turn' that takes an axis and an optional bool that reverses the rotation. Added some files that may be of use in the future, like rotation_pattern.h. Rotation patterns are still hard-coded into the turn_... functions.

+++ src/rubiks/colored_numbered_tile.h
@@ -9,7 +9,7 @@ namespace rubiks {
 class colored_numbered_tile : public colored_tile, public numbered_tile {
 public:
 	colored_numbered_tile() = default;
-	colored_numbered_tile(int face, int tile) :
+	colored_numbered_tile(size_t face, size_t tile) :
 		tile_base{face, tile},
 		colored_tile{face, tile},
 		numbered_tile{face, tile}
